add adapter-zone tests for dbt show
